Fix gba_cart.ld support for overlays
The current linker script uses `__iwram_overlay_end - __iwram_overlay_start` as an offset for `__ewram_lma`, which uses only the maximum size of an IWRAM overlay and is wrong even for only 2 overlays (section `.ewram`'s LMA will overlap with `.iwram1`'s LMA). This fix makes it correctly consider the total size of all overlays when computing the LMA for section `.ewram`.