Ensure main thread "tbss" aligned matches .tbss
SIZEOF(.tdata) doesn't account for the extra alignment added to the
start of `.tbss`, but TLS offsets do account for that. As a result, we
can end up with misaligned reads/writes for uninitialized thread-locals.

Additionally, the .tdata and .tbss sections were using the form:

    .section ALIGN(4) : { ... }

This forces the section to start at a 4-aligned address, even if the
section should have alignment greater than 4. By changing to the form

    .section : ALIGN(4) { ... }

We can ensure a _minimum_ alignment of 4, but the section will still be
aligned properly if it requires a higher alignment.